+++ qlit/server.py
@@ -10,7 +10,6 @@ THESAURUS = Thesaurus().parse('qlit.nt')
 print(f'Loaded thesaurus with {len(THESAURUS.refs())} terms')
 
 THESAURUS_SIMPLE = SimpleThesaurus() + THESAURUS
-THESAURUS_SIMPLE.rebuild(debug=True)
 
 FORMATS = {
     'ttl': 'text/turtle',
@@ -80,6 +79,8 @@ def api_labels():
 
 @app.route("/api/autocomplete")
 def api_autocomplete():
+    if not THESAURUS_SIMPLE.index:
+        THESAURUS_SIMPLE.rebuild(debug=True)
     # TODO Handle missing/bad arg
     s = request.args.get('s')
     return jsonify(THESAURUS_SIMPLE.autocomplete(s))