+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Changes to data are ignored for the purpose of this changelog.
 
 - MAJOR: Removed `/api/` route (`get_all`)
 - MAJOR: Renamed `autocomplete` to `search`, `parents` to `broader`, and `children` to `narrower`
-- Moved search mechanism to new module
+- MINOR: Rewrote search so no startup-time processing is needed
 - SimpleThesaurus does not _inherit_ Thesaurus, but _has_ it
 - Get labels directly from graph, not from SimpleTerm list